requirements.txt
requests==2.18.4
vcrpy==1.11.1
pytest==3.3.2